fix: split consolidated world tests to satisfy 400-line CI cap (#3843)
Unblock PR #3845 by splitting merged connectivity, fog, province lookup,
unit lookup, and province traversal test files at top-level group boundaries
while preserving shared test support helpers.
